+++ modules/cluster/cluster.go
@@ -4,6 +4,7 @@ import (
 	"github.com/infinitbyte/framework/core/cluster"
 	pb "github.com/infinitbyte/framework/core/cluster/pb"
 	"github.com/infinitbyte/framework/core/config"
+	"github.com/infinitbyte/framework/core/global"
 	"github.com/infinitbyte/framework/core/rpc"
 	"github.com/infinitbyte/framework/modules/cluster/demo/server"
 	"github.com/infinitbyte/framework/modules/cluster/discovery"
@@ -13,16 +14,22 @@ type ClusterModule struct {
 }
 
 func (module ClusterModule) Name() string {
-	return "ClusterName"
+	return "Cluster"
 }
 
 func (module ClusterModule) Setup(cfg *config.Config) {
+	if !global.Env().SystemConfig.ClusterConfig.Enabled {
+		return
+	}
 
-	rpc.Setup()
-	cluster.New()
+	rpc.Setup(&global.Env().SystemConfig.ClusterConfig.RPCConfig)
+	cluster.New(&global.Env().SystemConfig.ClusterConfig)
 }
 
 func (module ClusterModule) Start() error {
+	if !global.Env().SystemConfig.ClusterConfig.Enabled {
+		return nil
+	}
 
 	server.Init()
 
@@ -40,6 +47,10 @@ func (module ClusterModule) Start() error {
 }
 
 func (module ClusterModule) Stop() error {
+	if !global.Env().SystemConfig.ClusterConfig.Enabled {
+		return nil
+	}
+
 	cluster.SnapshotClusterState()
 	return nil
 }